refactor: move graphql.ts to a common folder
- move graphql.ts to a common folder

- update imports

- remove duplicate graphql.ts
