+++ src/packages/get-packages.ts
@@ -3,7 +3,7 @@ import {GraphQlQueryResponse} from '@octokit/graphql/dist-types/types'
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import {Observable, from, throwError} from 'rxjs'
 import {catchError, map} from 'rxjs/operators'
-import {graphql} from './graphql'
+import {graphql} from '../common/graphql'
 
 export interface PackageInfo {
   id: string
