+++ src/version/delete-version.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import {from, Observable, merge, throwError, of} from 'rxjs'
 import {catchError, map, tap} from 'rxjs/operators'
-import {graphql} from './graphql'
+import {graphql} from '../common/graphql'
 
 let deleted = 0
